+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-    - Issue 8 - Hydrocron API implementation with dynamodb local database
+    - Issue 8 - Hydrocron API implementation with mysql local database
 ### Changed 
+    - Issue 8 - Hydrocron API implementation with dynamodb local database
+    - Issue 8 - Rearrange database code
+    - Issue 8 - Rearrange tests
+    - Issue 8 - Add script to load data
 ### Deprecated 
 ### Removed
 ### Fixed